Completed switch to hydra
