+++ JobStreet/scrape.py
@@ -8,14 +8,11 @@ import os
 from pprint import pformat
 import time
 import datetime
-import argparse
+import hydra
 import random
 
-# Links
-jobstreet_main_link = 'http://www.jobstreet.com.sg'
-jobstreet_search_link = 'http://www.jobstreet.com.sg/en/job-search/'
-
 def return_timestamp():
+    """generates timestamp in YYYYMMDDhhmmss format"""
     dt_time = datetime.datetime.today()
     return str(10000000000*dt_time.year + 100000000*dt_time.month + 1000000*dt_time.day + 10000*dt_time.hour + 100*dt_time.minute + dt_time.second)
 
@@ -44,8 +41,20 @@ def log_decorator(log_name):
         return new_function
     return log_this
 
+def write_into_jsonl(path:str, job_listings:List[dict]):
+    """Writes the output into json file
+
+    Args:
+        path (str): output path
+        job_listings (List[dict]): all of the job info
+        overwrite (bool, optional): whether to overwrite in the output path. Defaults to False.
+    """
+    with open(path, 'w') as writer:
+        for job_listing in job_listings:
+            writer.write(json.dumps(job_listing) + '\n')
+
 @log_decorator(logger.name)
-def return_com_overview(results)->dict:
+def return_company_overview(results)->dict:
     """returns a dictionary filled with company overview information
 
     Args:
@@ -54,11 +63,11 @@ def return_com_overview(results)->dict:
     Returns:
         dict: Key: Company overview, value: text
     """
-    com_overview_dict = {}
+    company_overview_dict = {}
     for section in results:
         if str(section.text)[:16].lower() == 'company overview':
-            com_overview_dict['Company Overview'] = str(section.text)[16:]
-    return com_overview_dict
+            company_overview_dict['Company Overview'] = str(section.text)[16:]
+    return company_overview_dict
 
 @log_decorator(logger.name)
 def return_job_details(results)->dict:
@@ -76,9 +85,8 @@ def return_job_details(results)->dict:
     # its false by default coz in the first run I am always expecting a header first before lists will appear
     new_header = False
     temp_string = ""
-    for idx, content in enumerate(results[0].find_all(['li','strong'])):
-        if content.name == "strong":
-            # print(f"Header : {content.text}")
+    for idx, content in enumerate(results[0].find_all(['li','strong','b'])):
+        if content.name == "strong" or content.name == "b":
             if new_header == True:
                 last_key = list(temp_dict.keys())[-1]
                 temp_dict[last_key] = temp_string
@@ -88,12 +96,11 @@ def return_job_details(results)->dict:
             new_header = True
             continue
         if (content.name == 'li'):
-            # print(f"List content: {content.text}")
             if content.string is None:
                 temp_string = ""
             else:
                 temp_string += content.string + "."
-        if idx == len(results[0].find_all(['li','strong']))-1:
+        if idx == len(results[0].find_all(['li','strong','b']))-1:
             last_key = list(temp_dict.keys())[-1]
             temp_dict[last_key] = temp_string
 
@@ -101,50 +108,86 @@ def return_job_details(results)->dict:
 
 @log_decorator(logger.name)
 def return_company_name(results)->str:
+    """Returns the company name
+
+    Args:
+        results (_type_): _description_
+
+    Returns:
+        str: _description_
+    """
     company_name = results[0].string
     return company_name
 
 @log_decorator(logger.name)
 def return_small_section_text(results)->List:
+    """The small section is the upper left corner that contains details like
+    position, company name, address and date of posting
+
+    Args:
+        results (_type_): _description_
+
+    Returns:
+        List: _description_
+    """
     small_section_text = [data.string for data in results]
     return small_section_text
 
 @log_decorator(logger.name)
 def return_add_info(results)->dict:
+    """The additional information captured includes the 'additional information' under
+    the main JD and the 'additional company information' after company overview.
+    Returns everything in a dictionary with its content as value and header
+    as key 
+    """
     temp_dict = defaultdict(list)
     for item in results:
-        key = item.findAll('span')[0].text
-        value = item.findAll('span')[1].text
+        key = item.find_all('span')[0].text
+        value = item.find_all('span')[1].text
         temp_dict[key] = value
     return temp_dict
 
 @log_decorator(logger.name)
-def return_all_info(results_job_link):
+def return_all_info(results_job_link:str,params:dict)->dict:
+    """Main method that does all of the job and calls subordinate methods
+    for each section and returns all of the information contained
+
+    Args:
+        results_job_link (str): URL of one job 
+
+    Returns:
+        dict: results
+    """
     soup = bs4.BeautifulSoup(results_job_link.text,'html.parser')
-    small_section = soup.find_all('div', {"class": "sx2jih0 zcydq86a"})
-    company_name = soup.find_all('span', {"class": "sx2jih0 zcydq84u _18qlyvc0 _18qlyvc1x _18qlyvc2 _1d0g9qk4 _18qlyvcb"})
+    small_section = soup.find_all('div', {"class": params['small_section']})
+    company_name = soup.find_all('span', {"class": params['company_name']})
     jd = soup.select('[data-automation="jobDescription"]')
-    add_info = soup.find_all('div', {"class": "sx2jih0 zcydq8r _1tqfum90 _1tqfum97"})
-    com_overview = soup.findAll('div', {'class': "sx2jih0 zcydq86q zcydq86v zcydq86w"})
+    add_info = soup.find_all('div', {"class": params['add_info']})
+    company_overview = soup.find_all('div', {'class': params['company_overview']})
 
     small_section = return_small_section_text(small_section)
     company_name = return_company_name(company_name)
     jd = return_job_details(jd)
     add_info = return_add_info(add_info)
-    com_overview = return_com_overview(com_overview)
+    company_overview = return_company_overview(company_overview)
 
-    return small_section, company_name, jd, add_info, com_overview
+    return small_section, company_name, jd, add_info, company_overview
 
-def write_into_jsonl(path:str, job_listings:List[dict], overwrite:bool=False):
-    if os.path.exists(path) and overwrite == False:
-        logger.info(f'Unable to write output into {path} as path exists and overwrite is False')
-    else:
-        with open(path, 'w') as writer:
-            for job_listing in job_listings:
-                writer.write(json.dumps(job_listing) + '\n')
+
 
 @log_decorator(logger.name)
-def extract_from_one_job_listing(job_link, position)->dict:
+def extract_from_one_job_listing(job_link:str, position:str, params:dict)->dict:
+    """Returns all of the details from one job listing
+
+    Args:
+        job_link (str): url link for the job post
+        position (str): position
+
+    Returns:
+        dict: If no error returns position, company name, small section, url 
+        otherwise will return error message with position and URL
+    """
+    jobstreet_main_link = params['jobstreet_main_link']
     job_link = jobstreet_main_link+job_link
     results_job_link = requests.get(job_link)
 
@@ -152,8 +195,8 @@ def extract_from_one_job_listing(job_link, position)->dict:
         logger.info('Job street job link is Good to go!')
 
         try:
-            small_section, company_name, jd, add_info, com_overview = return_all_info(results_job_link)
-            return {"position":position},{"company_name":company_name},{"small_section":small_section},jd, add_info, com_overview, {"url":job_link}
+            small_section, company_name, jd, add_info, company_overview = return_all_info(results_job_link,params)
+            return {"position":position},{"company_name":company_name},{"small_section":small_section},jd, add_info, company_overview, {"url":job_link}
 
         except Exception as e:
             logger.error(f"Unable to extract info due to {e}")
@@ -162,7 +205,15 @@ def extract_from_one_job_listing(job_link, position)->dict:
         logger.error(f"Unable to extract job listing info from {job_link} for position {position}")
         return {'Error': {'job_link': job_link, 'Message':"status_code", 'position':position}}
 
-def extract_from_page(link:str)->List:
+def extract_from_page(link:str, params:dict)->List:
+    """Extracts all of the details in one page
+
+    Args:
+        link (str): link
+
+    Returns:
+        List: _description_
+    """
     job_listings = []
     results = requests.get(link)
     if results.status_code == 200:
@@ -179,20 +230,47 @@ def extract_from_page(link:str)->List:
         time.sleep(random.randint(0, 3))
         position = headers[idx].string
         job_link=headers[idx].get('href')
-        job_listings.append(extract_from_one_job_listing(job_link, position))
+        job_listings.append(extract_from_one_job_listing(job_link, position, params))
         
     return job_listings
 
-
+@hydra.main(config_path='./conf', config_name='hydra_config.yaml')
 def main(args):
+    """main function that takes in hydra configs and generates 
+    output
+    """
+
+    #############################################
+    #####  initialisation of variables      #####
+    #############################################
 
     job_listings = []
-    page_num = args.pages
-
+    jobstreet_main_link = args['settings']['jobstreet_main_link']
+    jobstreet_search_link = args['settings']['jobstreet_search_link']
+    page_num = args['settings']['pages']
     # the link needs the additional term of jobs in order for the url to be valid
-    search_term = args.search_term + '-jobs'
-
+    search_term = args['settings']['search_term'] + '-jobs'
     link = jobstreet_search_link + search_term
+
+    small_section_param = args['html_tag']['small_section']
+    company_name_param = args['html_tag']['company_name']
+    add_info_param = args['html_tag']['add_info']
+    company_overview_param = args['html_tag']['company_overview']
+
+    params = {
+        'small_section':small_section_param,
+        'company_name':company_name_param, 
+        'add_info':add_info_param, 
+        'company_overview':company_overview_param,
+        'jobstreet_main_link':jobstreet_main_link,
+        'jobstreet_search_link':jobstreet_search_link
+        }
+
+
+    #############################################
+    #####  searches page by page and extract#####
+    #############################################
+
 
     for pg_num in range(1,page_num+1):
         time.sleep(random.randint(0, 3))
@@ -200,9 +278,14 @@ def main(args):
         # jobstreet search link + search term + page num
         link = jobstreet_search_link+ search_term+"/"+ str(pg_num)+"/"
 
-        temp_listing = extract_from_page(link)
+        temp_listing = extract_from_page(link,params)
         for listing in temp_listing:
             job_listings.append(listing)
+
+
+    #############################################
+    #####           Writes file             #####
+    #############################################
 
     curr_dir = os.getcwd()
     sample_path = curr_dir+'/'+ search_term + timestamp +'.jsonl'
@@ -210,12 +293,4 @@ def main(args):
 
 
 if __name__ == "__main__":
-    parser =argparse.ArgumentParser()
-    parser.add_argument('--pages', type=int, default=2, help='Number of pages in the search term. \
-        If you entered a large number than what is available, the link will go back to the first page ')
-    parser.add_argument('--search_term', type=str, default='Data-scientist', help='The job that you want')
-
-    args = parser.parse_args()
-
-    logger.info(f" Search parameters are {args}")
-    main(args)
+    main()
